Ready to be tested i think

+++ aimbots-src/src/informants/robot-states/robot_state.cpp
@@ -1,7 +1,7 @@
 #include "robot_state.hpp"
 namespace src::RobotStates {
-Robot::Robot(float x, float y, float z, int robotId, int health, Team teamColor) {
-    this->position = Matrix<float, 3, 1>().zeroMatrix();
+Robot::Robot(short x, short y, short z, int robotId, int health, Team teamColor) {
+    this->position = Matrix<short, 3, 1>().zeroMatrix();
     position[Position::x][0] = x;
     position[Position::y][0] = y;
     position[Position::z][0] = z;
@@ -11,7 +11,7 @@ Robot::Robot(float x, float y, float z, int robotId, int health, Team teamColor)
     this->teamColor = teamColor;
 }
 
-Robot::Robot(Matrix<float, 3, 1> position, int robotId, int health, Team teamColor) {
+Robot::Robot(Matrix<short, 3, 1> position, int robotId, int health, Team teamColor) {
     this->position = position;
     this->robotID = robotID;
     this->health = health;
@@ -20,36 +20,36 @@ Robot::Robot(Matrix<float, 3, 1> position, int robotId, int health, Team teamCol
 }
 
 Robot::Robot() {
-    position = Matrix<float, 3, 1>().zeroMatrix();
+    position = Matrix<short, 3, 1>().zeroMatrix();
     robotID = -1;
     health = -1;
 }
 
 Robot::~Robot() {}
 
-void Robot::setX(float x) { this->position[Position::x][0] = x; }
+void Robot::setX(short x) { this->position[Position::x][0] = x; }
 
-void Robot::setY(float y) { this->position[Position::y][0] = y; }
+void Robot::setY(short y) { this->position[Position::y][0] = y; }
 
-void Robot::setZ(float z) { this->position[Position::z][0] = z; }
+void Robot::setZ(short z) { this->position[Position::z][0] = z; }
 
-void Robot::setPosition(float x, float y, float z) {
+void Robot::setPosition(short x, short y, short z) {
     position[Position::x][0] = x;
     position[Position::y][0] = y;
     position[Position::z][0] = z;
 }
 
-void Robot::setPosition(Matrix<float, 3, 1> position) { this->position = position; }
+void Robot::setPosition(Matrix<short, 3, 1> position) { this->position = position; }
 
 void Robot::setHealth(int health) { this->health = health; }
 
-float Robot::getX() { return this->position[Position::x][0]; }
+short Robot::getX() { return this->position[Position::x][0]; }
 
-float Robot::getY() { return this->position[Position::y][0]; }
+short Robot::getY() { return this->position[Position::y][0]; }
 
-float Robot::getZ() { return this->position[Position::z][0]; }
+short Robot::getZ() { return this->position[Position::z][0]; }
 
-Matrix<float, 3, 1> Robot::getPosition() { return this->position; }
+Matrix<short, 3, 1> Robot::getPosition() { return this->position; }
 
 int Robot::getHealth() { return this->health; }
